+++ src/main/java/net/pedroksl/advanced_ae/client/renderer/QuantumComputerInternalBakedModel.java
@@ -5,9 +5,13 @@ import java.util.HashMap;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.core.Direction;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.pedroksl.advanced_ae.common.blocks.AAECraftingUnitBlock;
 import net.pedroksl.advanced_ae.common.blocks.AAECraftingUnitType;
+import net.pedroksl.ae2addonlib.client.render.ConnectedTexturesBaseBakedModel;
 
-public class QuantumComputerInternalBakedModel extends QuantumComputerBaseBakedModel {
+public class QuantumComputerInternalBakedModel extends ConnectedTexturesBaseBakedModel {
     public QuantumComputerInternalBakedModel(
             TextureAtlasSprite face,
             TextureAtlasSprite side,
@@ -15,7 +19,7 @@ public class QuantumComputerInternalBakedModel extends QuantumComputerBaseBakedM
             TextureAtlasSprite faceAnimation,
             TextureAtlasSprite faceTopAnimation,
             TextureAtlasSprite faceBottomAnimation) {
-        super(RenderType.CUTOUT, face, side, poweredSides, block -> block.type != AAECraftingUnitType.STRUCTURE);
+        super(RenderType.CUTOUT, face, side, poweredSides);
 
         setSideEmissive(true);
 
@@ -27,5 +31,15 @@ public class QuantumComputerInternalBakedModel extends QuantumComputerBaseBakedM
         animationMap.put(Direction.SOUTH, faceAnimation);
         animationMap.put(Direction.EAST, faceAnimation);
         setFaceAnimation(animationMap, true);
+    }
+
+    @Override
+    protected boolean shouldConnect(Block block) {
+        return block instanceof AAECraftingUnitBlock unit && unit.type != AAECraftingUnitType.STRUCTURE;
+    }
+
+    @Override
+    protected boolean shouldBeEmissive(BlockState state) {
+        return state.getValue(AAECraftingUnitBlock.POWERED);
     }
 }