+++ src/main/java/net/pedroksl/advanced_ae/client/widgets/QuantumUpgradeWidget.java
@@ -12,9 +12,9 @@ import net.neoforged.api.distmarker.OnlyIn;
 import net.neoforged.neoforge.network.PacketDistributor;
 import net.pedroksl.advanced_ae.client.gui.QuantumArmorConfigScreen;
 import net.pedroksl.advanced_ae.client.gui.widgets.AAEIcon;
-import net.pedroksl.advanced_ae.client.gui.widgets.AAEIconButton;
 import net.pedroksl.advanced_ae.common.items.upgrades.UpgradeType;
 import net.pedroksl.advanced_ae.network.packet.quantumarmor.QuantumArmorUpgradeTogglePacket;
+import net.pedroksl.ae2addonlib.client.widgets.AddonIconButton;
 
 import appeng.client.gui.Icon;
 import appeng.client.gui.style.ScreenStyle;
@@ -138,7 +138,7 @@ public class QuantumUpgradeWidget {
         }
     }
 
-    static class UninstallButton extends AAEIconButton {
+    static class UninstallButton extends AddonIconButton {
         public UninstallButton(int x, int y, OnPress onPress) {
             super(onPress);
             setX(x);