+++ src/main/java/net/pedroksl/advanced_ae/common/definitions/AAEText.java
@@ -86,7 +86,7 @@ public enum AAEText implements LocalizationEnum {
     FilteredImportOn("Input must match crafting outputs while crafting jobs are active", Type.TOOLTIP),
     FilteredImportOff("Input is unrestricted", Type.TOOLTIP),
     DirectionalOutput("Directional Output", Type.TOOLTIP),
-    DirectionOutputHint("Configure which directions are allowed for output auto-export", Type.TOOLTIP),
+    DirectionalOutputHint("Configure which directions are allowed for output auto-export", Type.TOOLTIP),
 
     NumberTextFieldInputHint("Confirm the input using the %s key.", Type.TOOLTIP),
     NumberTextFieldOutputHint(
@@ -96,7 +96,9 @@ public enum AAEText implements LocalizationEnum {
     EnablePatternButton("Enabled/Disable", Type.TOOLTIP),
 
     SetAmountButtonHint("Middle click to set the amount to keep in stock", Type.GUI),
-    SetAmount("Set Stock Amount", Type.GUI),
+    SetAmount("Set Amount", Type.GUI),
+    Enabled("Enabled", Type.TOOLTIP),
+    Disabled("Disabled", Type.TOOLTIP),
 
     ThroughputMonitorValue("%s" + "%s" + "/s", Type.GUI),
     OverdriveThroughputMonitorValue("%s" + "%s" + "/t", Type.GUI),