+++ src/main/java/net/pedroksl/advanced_ae/gui/StockExportBusMenu.java
@@ -5,9 +5,10 @@ import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.inventory.Slot;
-import net.pedroksl.advanced_ae.api.ISetAmountMenuHost;
 import net.pedroksl.advanced_ae.common.definitions.AAEMenus;
 import net.pedroksl.advanced_ae.common.parts.StockExportBusPart;
+import net.pedroksl.ae2addonlib.api.ISetAmountMenuHost;
+import net.pedroksl.ae2addonlib.gui.SetAmountMenu;
 
 import appeng.api.stacks.GenericStack;
 import appeng.core.definitions.AEItems;